requirements.txt
python-dotenv==0.15.0
flask>=1.0,<=1.1.2
pytest==6.2.2
Flask-Mail==0.9.1
psycopg2-binary==2.8.6